requirements.txt
streamlit
python-mcp
mistralai
httpx
python-dotenv
